app/requirements.txt
fastapi
uvicorn
prometheus-client
python-jose[cryptography]
pydantic